Create 99999-battleships.py
Jack plays a game of battleships with his friend Stacy. The game is played on a square map of N rows, numbered from 1 to N. Each row contains N cells, labeled with consecutive English upper-case letters (A, B, C, etc.). Each cell is identified by a string composed of its row number followed by its column number: for example, "9C" denotes the third cell in the 9th row, and "15D" denotes the fourth cell in the 15th row. Jack marks the positions of all his ships on the map (which is not shown to Stacy). Ships are defined by rectangles with a maximum area of 4 cells. Stacy picks map cells to hit some ships. A ship is considered to be hit if at least one of its constituent cells is hit. If all of a ship's cells are hit, the ship is sunk. The goal is to count the number of sunk ships and the number of ships that have been hit but not
sunk.